fix(login): Start migrating login to use auth0-react.
Fix #335
