+++ lib/common/components/Login.js
@@ -1,28 +1,30 @@
 // @flow
 
-import {Component} from 'react'
+import { Component } from 'react'
 
-import * as userActions from '../../manager/actions/user'
-import auth0 from '../user/Auth0Manager'
-
-import type {Props as ContainerProps} from '../containers/Login'
+import type { Props as ContainerProps } from '../containers/Login'
 
 type Props = ContainerProps & {
-  onHide: () => void,
-  push: string => void,
-  receiveTokenAndProfile: typeof userActions.receiveTokenAndProfile,
-  redirectOnSuccess: string
+  loggedIn: boolean,
+  push: string => void
 }
 
-export default class Auth0 extends Component<Props> {
+export default class Login extends Component<Props> {
   componentDidMount () {
-    // Show lock when component mounts. NOTE: render method returns null.
-    auth0.loginWithLock(this.props)
+    const { auth0, loggedIn, query, redirectOnSuccess } = this.props
+    if (!query.to && !loggedIn && redirectOnSuccess) {
+      // Show lock when component mounts. NOTE: render method returns null.
+      auth0.loginWithRedirect({
+        redirectUri: `${window.location.origin}/login?to=${redirectOnSuccess}`
+      })
+    }
   }
 
-  componentWillUnmount () {
-    // Hide lock when component unmounts
-    auth0.hideLock()
+  componentDidUpdate () {
+    const { loggedIn, push, query } = this.props
+    if (query.to && loggedIn) {
+      push(query.to)
+    }
   }
 
   render () {
