+++ lib/common/containers/Login.js
@@ -1,33 +1,32 @@
 // @flow
 
-import {connect} from 'react-redux'
-import {push} from 'react-router-redux/lib/actions'
+import { Auth0ContextInterface, withAuth0 } from '@auth0/auth0-react'
+import { connect } from 'react-redux'
+import { push } from 'react-router-redux/lib/actions'
 
 import Login from '../components/Login'
-import {receiveTokenAndProfile} from '../../manager/actions/user'
-
-import type {AppState, RouterProps} from '../../types/reducers'
+import type { AppState, RouterProps } from '../../types/reducers'
 
 export type Props = $Shape<RouterProps> & {
+  auth0: Auth0ContextInterface,
   onHide?: () => void,
+  query: any,
   redirectOnSuccess?: string
 }
 
 const mapStateToProps = (state: AppState, ownProps: Props) => {
-  const {redirectOnSuccess} = state.user
+  const { redirectOnSuccess } = state.user
+  const { locationBeforeTransitions } = state.routing
   return {
+    loggedIn: ownProps.auth0.isAuthenticated,
+    query: (locationBeforeTransitions && locationBeforeTransitions.query) || {},
     // If redirect URL is null, we want to default to application home.
     redirectOnSuccess: redirectOnSuccess || window.location.path || '/home'
   }
 }
 
 const mapDispatchToProps = {
-  push,
-  receiveTokenAndProfile,
-  // Default onHide action is to go back to root page
-  onHide: () => push('/')
+  push
 }
 
-// this was done because there was an error I couldn't figure out in StatusModal
-const connected: any = connect(mapStateToProps, mapDispatchToProps)(Login)
-export default connected
+export default withAuth0(connect(mapStateToProps, mapDispatchToProps)(Login))
