+++ lib/common/user/Auth0Manager.js
@@ -9,10 +9,10 @@ import uuidv4 from 'uuid/v4'
 import * as userActions from '../../manager/actions/user'
 import {getComponentMessages, getConfigProperty} from '../util/config'
 
-const clientID = process.env.AUTH0_CLIENT_ID || ''
-const domain = process.env.AUTH0_DOMAIN || ''
-const auth0Client = new auth0.WebAuth({ domain, clientID })
-const DEFAULT_SCOPE = 'app_metadata profile email openid user_metadata'
+export const clientID = process.env.AUTH0_CLIENT_ID || ''
+export const domain = process.env.AUTH0_DOMAIN || ''
+export const auth0Client = new auth0.WebAuth({ domain, clientID })
+export const DEFAULT_SCOPE = 'app_metadata profile email openid user_metadata'
 
 class Auth0Manager {
   isTryingToGetProfileFromToken: boolean
